requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-chess==1.999
stockfish==3.28.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.7
python-chess==1.999
sqlalchemy==2.0.36
psycopg2-binary==2.9.7

